fix:(PersistentUnorderedMap) JSDoc documentation fix (#234)
* fix:(PersistentUnorderedMap) JSDoc documentation fix
* fix:(PersistentUnorderedMap) faster delete
