+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules
 \.idea/
+\.vs/
 
 *.wasm
