+++ assembly/sdk/collections/persistentUnorderedMap.ts
@@ -7,13 +7,13 @@ export class PersistentUnorderedMap<K, V> {
   private _entries: PersistentVector<MapEntry<K, V>>;
 
   /**
-   * Creates or restores a persistent map with a given storage prefix.
+   * Creates or restores a persistent unordered map with a given storage prefix.
    * Always use a unique storage prefix for different collections.
    *
    * Example
    *
    * ```ts
-   * let map = new PersistentMap<string, string>("m") // note the prefix must be unique (per NEAR account)
+   * let map = new PersistentUnorderedMap<string, string>("m") // note the prefix must be unique (per NEAR account)
    * ```
    *
    * @param prefix A prefix to use for every key of this map.
@@ -27,7 +27,7 @@ export class PersistentUnorderedMap<K, V> {
    * Retrieves the related value for a given key, or uses the `defaultValue` if not key is found
    *
    * ```ts
-   * let map = new PersistentMap<string, string>("m")
+   * let map = new PersistentUnorderedMap<string, string>("m")
    *
    * map.set("hello", "world")
    * let found = map.get("hello")
@@ -52,7 +52,7 @@ export class PersistentUnorderedMap<K, V> {
    * Retrieves a related value for a given key or fails assertion with "key not found"
    *
    * ```ts
-   * let map = new PersistentMap<string, string>("m")
+   * let map = new PersistentUnorderedMap<string, string>("m")
    *
    * map.set("hello", "world")
    * let result = map.getSome("hello")
@@ -62,7 +62,7 @@ export class PersistentUnorderedMap<K, V> {
    * ```
    *
    * @param key Key of the element.
-   * @returns Value for the given key or the default value.
+   * @returns Value for the given key or throws
    */
   getSome(key: K): V {
     let index = this._map.getSome(key);
@@ -70,7 +70,7 @@ export class PersistentUnorderedMap<K, V> {
   }
 
   /**
-   * returns a map of number of last entries.
+   * returns a MapEntry array of number of last entries.
    * @param count
    */
   last(count: i32): MapEntry<K, V>[] {
@@ -88,7 +88,7 @@ export class PersistentUnorderedMap<K, V> {
    * Checks whether the map contains a given key
    *
    * ```ts
-   * let map = new PersistentMap<string, string>("m")
+   * let map = new PersistentUnorderedMap<string, string>("m")
    *
    * map.contains("hello")      // false
    * map.set("hello", "world")
@@ -160,7 +160,7 @@ export class PersistentUnorderedMap<K, V> {
 
   /**
    * ```ts
-   * let map = new PersistentMap<string, string>("m")
+   * let map = new PersistentUnorderedMap<string, string>("m")
    *
    * map.set("hello", "world")
    * ```
@@ -186,7 +186,7 @@ export class PersistentUnorderedMap<K, V> {
    * Removes the given key and related value from the map
    *
    * ```ts
-   * let map = new PersistentMap<string, string>("m")
+   * let map = new PersistentUnorderedMap<string, string>("m")
    *
    * map.set("hello", "world")
    * map.delete("hello")
@@ -196,23 +196,28 @@ export class PersistentUnorderedMap<K, V> {
    * @param key Key to remove.
    */
   delete(key: K): void {
-    // make sure the item is in the set
-    assert(this._map.contains(key), "The item was not found in the set");
+    // get index and also make sure the item is in the set
+    const index = this._map.getSome(key);
 
-    // swap_remove requires at least 2 elements to work so a single element
-    // is the same as clearing
-    if (this._entries.length == 1) {
-      this.clear();
-      return;
+    if (index == this._entries.length - 1) {
+      //deleting the last entry (or the only element left)
+      //we avoid using swap_remove in order to use fewer operations to remove the entry
+      let entry = this._entries.popBack(); //remove the last entry
+      this._map.delete(entry.key); //remove the key=>index map
+      return; //early exit
     }
 
-    // remove the item from the set
-    const swapKey = this._entries.last.key;
-    const index = this._map.getSome(key);
-    this._entries.swap_remove(index);
+    //Here we're removing a key in the middle [0..length-2]
+    //we'll use swap_remove, which removes the item from the vec
+    //by placing the last item in its position (O(1))
 
-    // update our accounting of items in the set
-    this._map.set(swapKey, index);
+    // get the key of the last item in the vec
+    const lastKey = this._entries.last.key;
+    // remove the entry from the vec by placing the last item in its position (swap_remove)
+    this._entries.swap_remove(index);
+    // update the new position of the now moved lastKey
+    this._map.set(lastKey, index);
+    // delete the key from the key=>index map
     this._map.delete(key);
   }
